requirements.txt
flask
openai
python-dotenv
python-docx 
fpdf2
gunicorn